src: Rename several X25519 functions.
The original naming was intended to prevent vulnerabilities from failing to include the public keys in key derivation from a shared secret. However, ComputeXCoordinate() sounds horrible, and DeriveSharedSecret() is misleading when it's doing more than that. Given BLAKE2bHashAlgorithm has been removed, it seems appropriate to adjust this now as well.