src: Support caching the state with IncrementalBLAKE2b.
Passing the state via ref, like in NSec/Monocypher.NET, allows you to cache the state after processing static data like a key. However, switching to that style of API would be a breaking change and is dodgy with algorithms like Poly1305 where you shouldn't be able to cache the state. Therefore, I've opted for this approach. It only allows caching once so the cached state can be zeroed, but that should be sufficient in most cases (e.g., a key or a key plus some other static data).