SecureRandom.cs: Fix GetPassphrase() bug in .NET 8.
Whilst using Geralt in a .NET 8 CLI program, I noticed that this method returned passphrases with '\r' characters in them on Windows, resulting in them printing to the console incorrectly. This was not an issue in .NET 6 and isn't detected in the tests, even with the changes to the tests. Therefore, there's been some change between .NET 6-8 that affects line endings for embedded files but not in a library context... Regardless, this change is how it should've been written originally for safety.