src: Add guarded heap allocations.
I presume MaxSize is appropriate, but the libsodium code is a little confusing as the pointers make it look like the data and canary don't necessarily go in the same page. It's not great this is potentially variable as well.

AsSpan() is a bit dangerous because it can cause an access violation after Dispose() if you're not careful. It's best to call it each time rather than creating a span variable.

NoAccess(), ReadOnly(), and ReadWrite() might want to be renamed. For example, adding 'Access' to the end of each or adding 'Set' to the beginning of each.

sodium_allocarray() isn't supported because you can just slice the returned span. It just calls sodium_malloc() anyway.