src: Check for disposed in all incremental classes.
This stuff isn't thread safe, but official .NET IDisposable APIs don't seem to be either. I might change that in the future.

The other thing to note is that it's unnecessary for the tests to have DynamicData. The same is true for InvalidOperation tests. It just inflates the number of tests, so I might change this.