SecureMemory.cs: Remove the memory locking functions.
These APIs are a bit dangerous because the buffer could be moved around in memory by .NET and memory locking operates on pages. I attempted to implement a LockedAllocation IDisposable class, but it appears that page boundary alignment cannot be guaranteed without calling OS APIs, which would be doable but is going out of scope. Instead, the GuardedHeapAllocation class can be used as it does memory locking for you. However, it allocates significantly more memory.