SecureRandom.cs: Change the string/passphrase methods to buffers.
This is more consistent with the rest of the library and helps with zeroing memory. However, it's messy for passphrases because the output length isn't known, so you have to slice the output.

This commit also adds improved validation, like for the upperBound in GetInt32() and null strings. Then there are a few code/test improvements.